requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI/ML
langgraph>=0.0.15
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# Database & Storage
chromadb>=0.4.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0

# Development
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0